Added message to admin use admin area to create items.
